requirements.txt
streamlit
tensorflow
numpy
pillow
matplotlib
pandas
plotly
gdown
